fix: rename size enum
