+++ src/components/CircleIconBtn/CircleIconBtnProps.ts
@@ -1,4 +1,4 @@
-import { Size } from "./index";
+import { SizeCircleIconBtn as Size } from "./index";
 
 export interface MetCircleIconBtnProps {
     /**
@@ -16,7 +16,7 @@ export interface MetCircleIconBtnProps {
     /**
      * Size of Button
      */
-    size?: Size.lg | Size.sm;
+    size?: Size.LAGER | Size.SMALL;
     /**
      * Function for trigger event onClick icon btn
      */
