--- src/components/CircleIconBtn/size.ts
@@ -1,4 +1,0 @@
-export enum Size {
-  lg = "lg",
-  sm = "sm",
-}
