+++ src/components/CircleIconBtn/sizeCircleIconBtn.ts
@@ -1,0 +1,4 @@
+export enum SizeCircleIconBtn {
+  LAGER,
+  SMALL
+}
